Move miner's logic from MainMenu to BaseScene
